requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic>=2.7.0,<3.0.0
langchain==0.2.15
groq==0.11.0
langchain-groq==0.1.9
python-dotenv==1.0.1
pytest==8.3.2

